requirements.txt
Django==1.8.4
PyJWT==1.4.0
appnope==0.1.0
decorator==4.0.2
defusedxml==0.4.1
djangorestframework==3.2.2
gnureadline==6.3.3
ipython==4.0.0
ipython-genutils==0.1.0
oauthlib==1.0.3
path.py==7.6
pexpect==3.3
pickleshare==0.5
python-social-auth==0.2.12
python3-openid==3.0.6
requests==2.7.0
requests-oauthlib==0.5.0
simplegeneric==0.8.1
six==1.9.0
traitlets==4.0.0
